+++ src/main/java/com/space/server/common/config/SecurityConfig.java
@@ -9,8 +9,8 @@ import com.space.server.common.jwt.filter.CustomLogoutFilter;
 import com.space.server.common.jwt.filter.CustomJwtFilter;
 import com.space.server.common.jwt.filter.LoginFilter;
 import com.space.server.common.jwt.util.JwtUtil;
-import com.space.server.common.oauth.handler.CustomSuccessHandler;
-import com.space.server.common.oauth.service.CustomOAuth2UserService;
+import com.space.server.domain.oauth.handler.CustomSuccessHandler;
+import com.space.server.domain.oauth.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
