+++ src/main/java/com/space/server/domain/oauth/service/CustomOAuth2UserService.java
@@ -1,6 +1,7 @@
-package com.space.server.common.oauth.service;
+package com.space.server.domain.oauth.service;
 
 import com.space.server.common.oauth.service.dto.*;
+import com.space.server.domain.oauth.service.dto.*;
 import com.space.server.domain.user.domain.Users;
 import com.space.server.domain.user.domain.repository.UserRepository;
 import com.space.server.domain.user.domain.value.Role;
