+++ src/main/java/com/space/server/domain/oauth/service/dto/GoogleResponse.java
@@ -1,23 +1,22 @@
-package com.space.server.common.oauth.service.dto;
+package com.space.server.domain.oauth.service.dto;
+
+import lombok.RequiredArgsConstructor;
 
 import java.util.Map;
 
-public class NaverResponse implements OAuth2Response {
+@RequiredArgsConstructor
+public class GoogleResponse implements OAuth2Response{
 
     private final Map<String, Object> attribute;
 
-    public NaverResponse(Map<String, Object> attribute) {
-        this.attribute = (Map<String, Object>) attribute.get("response");
-    }
-
     @Override
     public String getProvider() {
-        return "naver";
+        return "google";
     }
 
     @Override
     public String getProviderId() {
-        return attribute.get("id").toString();
+        return attribute.get("sub").toString();
     }
 
     @Override
