+++ src/main/java/com/space/server/domain/oauth/service/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.space.server.common.oauth.service.dto;
+package com.space.server.domain.oauth.service.dto;
 
 import com.space.server.domain.user.domain.value.Role;
 import lombok.Builder;
